requirements.txt
google-api-python-client
oauth2client
iptcinfo
numpy
Pillow
mock
